+++ sdk/test-utils/recorder/src/matcher.ts
@@ -10,16 +10,28 @@ export type Matcher = "HeaderlessMatcher" | "BodilessMatcher" | "CustomDefaultMa
 /**
  * Body the customer matcher expects.
  */
-export interface CustomMatcherBody {
+export interface CustomMatcherOptions {
   /**
    * Should the body value be compared during lookup operations?
    */
   compareBodies?: boolean;
   /**
    * Array of additional headers that should be excluded during matching.
-   * "Excluded" headers are entirely ignored. Unlike "ignored" headers, the presence (or lack of presence) of a header will not cause mismatch.
    */
   excludedHeaders?: string[];
+  /**
+   * By default, the test-proxy does not sort query params before matching. Setting true will sort query params alphabetically before comparing URI.
+   */
+  ignoreQueryOrdering?: boolean;
+}
+
+/**
+ * Body the customer matcher expects.
+ *
+ * // Ignored Headers option is not exposed to the users as of now.
+ * // If needed, this can be moved into CustomMatcherOptions.
+ */
+interface InternalCustomMatcherOptions extends CustomMatcherOptions {
   /**
    * Array of additional headers that should be ignored during matching.
    * Any headers that are "ignored" will not do value comparison when matching.
@@ -28,10 +40,6 @@ export interface CustomMatcherBody {
    * This also applies if the header is present in the request but not recording.
    */
   ignoredHeaders?: string[];
-  /**
-   * By default, the test-proxy does not sort query params before matching. Setting true will sort query params alphabetically before comparing URI.
-   */
-  ignoreQueryOrdering?: boolean;
 }
 
 export async function setMatcher(
@@ -39,7 +47,7 @@ export async function setMatcher(
   httpClient: HttpClient,
   matcher: Matcher,
   recordingId?: string,
-  matcherBody: CustomMatcherBody = { compareBodies: true, ignoreQueryOrdering: false }
+  matcherBody: InternalCustomMatcherOptions = { compareBodies: true, ignoreQueryOrdering: false }
 ): Promise<void> {
   const url = `${recorderUrl}${paths.admin}${paths.setMatcher}`;